backend/requirements.txt
Flask
Flask-MySQLdb
Flask-Cors
PyMySQL
Flask-Bcrypt
PyJWT
